+++ RELEASE_NOTES.md
@@ -1,5 +1,37 @@
 # Release Notes
 
+## Version 11.22.1
+_1 November 2021_
+
+### Flagged Features
+* Enable view/edit of Conformance Methods
+
+### Bug Fixes
+* Display added Optional Standards in Listing Report
+
+### Flagged Bug Fixes
+* Enable progress through confirmation wizard ignoring "warning" checkbox
+
+---
+
+## Version 11.22.0
+_18 October 2021_
+
+### Features
+* Disallow editing of Certification Date directly
+
+### User Interface Updates
+* Update text describing API Documentation
+
+### Bug Fixes
+* Give user an error message when saving Developer with bad URL
+* Require at least one Certification Status history item on edit
+
+### Flagged Features
+* Require user to attest before submitting Attestations
+
+---
+
 ## Version 11.21.0
 _5 October 2021_